+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.1.1
+
+### Added
+
+- LifetimeManagement extensions for base Addressables
+- invoking action when operation with addressables is complete
+- support for adding auto release trigger by asset string key
+
 ## v0.1.0
 
 ### Added